+++ src/onegov.core/HISTORY.rst
@@ -1,5 +1,9 @@
 Changelog
 ---------
+
+- Adds compatibility with Morepath 0.15.
+  [href]
+
 0.22.0 (2016-07-14)
 ~~~~~~~~~~~~~~~~~~~